+++ protocol/xmrtaker/swap_state_test.go
@@ -322,7 +322,7 @@ func TestSwapState_NotifyClaimed(t *testing.T) {
 	xmrAddr := kp.Address(common.Mainnet)
 
 	// lock xmr
-	tResp, err := maker.Transfer(xmrAddr, 0, uint(amt))
+	tResp, err := maker.Transfer(xmrAddr, 0, uint64(amt))
 	require.NoError(t, err)
 	t.Logf("transferred %d pico XMR (fees %d) to account %s", tResp.Amount, tResp.Fee, xmrAddr)
 	require.Equal(t, uint(amt), tResp.Amount)
